requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.18
python-dotenv==1.0.1
google-genai==0.3.0
Pillow==11.0.0
